Power-ups added, looking weak functions added, end game functionality added
If the player with rockets gets below a certain health he gets nerfed by having decreased movement speed, however he generates an ally fighter jet that he can shoot at to gain health back and regain movement speed. Additionally if a characters health drops below 10 they lose, the number is set at 10 so that there is a margin for bullets and rockets to stop doing damage before the game crashes, and allows for players to see which character died. If a character loses they will now become a floating you lose icon and lose the ability to shoot until the game is restarted